Declare trace types generically: ribbon first
This will help insulate me from my choice of plotting library.  I can
declare my plots generically, and only the specific implementation knows
how to turn that generic declaration into an implementation!

I'm starting with only ribbon plots as a proof of concept.

This introduces the language of "aes", or "aesthetics", which I'm
borrowing from R's ggplot2.  Maybe someday, I'll make the library more
compatible with the ideas of ggplot2 and the grammar of graphics.  But
for now, this is mainly aspirational.  (It probably also means that in
the meantime, I'm creating a confusing, half-assed ggplot2 wannabe.  Oh
well.)