Make augmentations generic by factoring out maps
Initially, I assumed an augmentation is something that transforms every
element of one dataset into another element.  This is not general
enough: I also want to be able to transform an arbitrary number of
points into a different arbitrary number of points, say, via GPs.

Fortunately, it was easy to factor out the generic idea of
"augmentations" from the more specific idea of augmentations that apply
to individual elements---"map augmentations," essentially.